+++ probe-rs/src/debug/stepping_mode.rs
@@ -1,7 +1,6 @@
 use super::debug_info::DebugInfo;
-use super::DebugError;
-use crate::core::Core;
-use crate::CoreStatus;
+use super::{halt_locations::HaltLocations, DebugError};
+use crate::{core::Core, CoreStatus};
 
 /// Stepping granularity for stepping through a program during debug.
 #[derive(Debug)]
@@ -10,9 +9,7 @@ pub enum SteppingMode {
     StepInstruction,
     /// Step Over the current statement, and halt at the start of the next statement.
     OverStatement,
-    /// DWARF doesn't encode when a statement contains a call to a non-inlined function.
-    /// - The best-effort approach is to step a single instruction, and then find the first valid breakpoint address.
-    /// - The worst case is that the user might have to perform the step into action more than once before the debugger properly steps into a function at the specified address.
+    /// Use best efforts to determin the location of any function calls in this statement, and step into them.
     IntoStatement,
     /// Step to the calling statement, immediately after the current function returns.
     OutOfStatement,
@@ -27,7 +24,7 @@ impl SteppingMode {
     /// - If no hardware breakpoints are available, we will do repeated instruction steps until we reach the desired location.
     ///
     /// Usage Note:
-    /// - Currently, no special provision is made for the effect of user defined breakpoints in interrupts that get triggered before this function completes.
+    /// - Currently, no special provision is made for the effect of interrupts that get triggered during stepping. The user must ensure that interrupts are disabled during stepping, or accept that stepping may be diverted by the interrupt processing on the core.
     pub fn step(
         &self,
         core: &mut Core<'_>,
@@ -57,17 +54,76 @@ impl SteppingMode {
         }
 
         let mut target_address: Option<u64> = None;
+        let mut adjusted_program_counter = program_counter;
+
         // Sometimes the target program_counter is at a location where the debug_info program row data does not contain valid statements for halt points.
         // When DebugError::NoValidHaltLocation happens, we will step to the next instruction and try again(until we can reasonably expect to have passed out of an epilogue), before giving up.
         for _ in 0..10 {
-            match super::halt_locations::HaltLocations::new(
-                debug_info,
-                program_counter,
-                Some(return_address),
-            ) {
+            match HaltLocations::new(debug_info, program_counter, Some(return_address)) {
                 Ok(program_row_data) => {
                     match self {
-                        SteppingMode::OverStatement | SteppingMode::IntoStatement => {
+                        SteppingMode::IntoStatement => {
+                            // This is a tricky case, for a couple of reasons:
+                            // - Firstly, the current RUST generated DWARF, does not store the DW_TAG_call_site information described in the DWARF 5 standard. It is not a mandatory attribute, so not sure if we can ever expect it.
+                            // - Secondly, because we want the first_halt_address to be based on the instructions of the called function, while next_halt_address and step_out_address must be based on the current sequence of statements.
+
+                            // To find (if they exist) functions called from the current program counter:
+                            // - At the current PC, determine the `next_statement_address` in the current sequence of instructions are.
+                            // - Single step the target core, until either ...
+                            //   (a) We hit a PC that is not in the sequence between starting PC and the address of the `next_statement_address` stored above. Halt at this location.
+                            //      (a.i) This could mean either that we encountered a branch (call to another instruction), or an interrupt handler diverted the processing.
+                            //   (b) The new PC matches the next valid statement stored above, which means there was nothing to step into, so the target is now halted (correctly) at the `next_halt_address`
+
+                            target_address = if let (
+                                Some(first_halt_location),
+                                Some(next_statement_address),
+                            ) = (
+                                program_row_data.first_halt_address,
+                                program_row_data.next_statement_address,
+                            ) {
+                                adjusted_program_counter = loop {
+                                    let next_pc = core.step()?.pc;
+                                    if (first_halt_location..next_statement_address)
+                                        .contains(&next_pc)
+                                    {
+                                        // We are still in the current sequence of instructions, before the next_statement_address.
+                                        continue;
+                                    } else if next_pc == next_statement_address {
+                                        // We have reached the next_statement_address, so we can conclude there was no branching calls in this sequence.
+                                        log::warn!("Stepping into next statement, but no branching calls found. Stepped to next available statement.");
+                                        break next_pc;
+                                    } else {
+                                        // We have reached a location that is not in the current sequence, so we can conclude there was a branching call in this sequence.
+                                        if let Some(valid_halt_address) =
+                                            HaltLocations::new(debug_info, next_pc, None)
+                                                .ok()
+                                                .and_then(|program_row_data| {
+                                                    program_row_data.next_statement_address
+                                                })
+                                        {
+                                            break valid_halt_address;
+                                        } else {
+                                            break next_pc;
+                                        }
+                                    }
+                                };
+                                Some(adjusted_program_counter)
+                            } else {
+                                // Our technique requires a valid first_halt_address AND a valid next_statement_address, so if we don't have one, we will later on step a single instruction.
+                                None
+                            };
+
+                            if target_address.is_none() {
+                                log::error!(
+                                    "Unable to determine target functions for stepping into instructions at {:x}. Stepping to next target instruction.",
+                                    program_counter
+                                );
+                                program_counter = core.step()?.pc;
+                                core_status = core.status()?;
+                                return Ok((core_status, program_counter));
+                            }
+                        }
+                        SteppingMode::OverStatement => {
                             target_address = program_row_data.next_statement_address
                         }
                         SteppingMode::OutOfStatement => {
@@ -116,9 +172,8 @@ impl SteppingMode {
                     target_address
                 );
 
-                if target_address == program_counter as u64 {
-                    // For simple functions that complete in a single statement.
-                    program_counter = core.step()?.pc;
+                if target_address == adjusted_program_counter as u64 {
+                    // For inline functions we have already stepped to the correct target address..
                 } else if core.set_hw_breakpoint(target_address).is_ok() {
                     core.run()?;
                     core.clear_hw_breakpoint(target_address)?;
@@ -126,7 +181,7 @@ impl SteppingMode {
                         Ok(core_status) => {
                             match core_status {
                                 CoreStatus::Halted(_) => {
-                                    program_counter =
+                                    adjusted_program_counter =
                                         core.read_core_reg(core.registers().program_counter())?
                                 }
                                 other => {
@@ -134,7 +189,7 @@ impl SteppingMode {
                                         "Core should be halted after stepping but is: {:?}",
                                         &other
                                     );
-                                    program_counter = 0;
+                                    adjusted_program_counter = 0;
                                 }
                             };
                             core_status
@@ -147,7 +202,7 @@ impl SteppingMode {
                         // TODO: In theory, this could go on for a long time. Should we consider NOT allowing this kind of stepping if there are no breakpoints available?
                     }
                     core_status = core.status()?;
-                    program_counter = target_address;
+                    adjusted_program_counter = target_address;
                 }
             }
             None => {
@@ -158,6 +213,6 @@ impl SteppingMode {
                 });
             }
         }
-        Ok((core_status, program_counter))
+        Ok((core_status, adjusted_program_counter))
     }
 }